FlappyBird.Game (FlappyBird v1.0.0)

Context contenant les fonctions pour le déroulement d'une Game.

Functions

flap(game)

Action de voler lorsque l'utilisateur appuis sur une touche ou clique sur l'écran de jeu.

Description

Reçois un game, puis modifie la variable velocity de son composant bird pour la valeur 3.

Exemples

iex> flap(%{field: value})
{:ok, %Game{}}

iex> flap(%{field: bad_value})
{:error, %Ecto.Changeset{}}

Type d'appel

Game.flap(game)

Type de retour

retourne game

new()

Création d'une nouvelle partie.

Exemples

iex> new()
{:ok, %Game{}}

Type d'appel

Game.new()

Valeur par défaut

game {
  bird: %{ x: 10, y: 35, wings: 0, velocity: 0 }
  pipes: []
  score: 0.0
}

save_score(v)

Sauvegarder le score d'une partie.

Variable d'argument

v -> value --> valeur du score à sauvegarder en string

Description

Appel la fonction Score.changeset() pour sauvegarder le score dans la base de données avec comme paramètre Entité/Structure Score vide --> %Score{} et la valeur a changer --> %{"value" => v}

Exemples

iex> save_score(value)
{:ok, %Game{}}

iex> save_score(bad_value)
{:error, %Ecto.Changeset{}}

update(game)

Mise à jour de la partie. Boucle d'exécution principale du jeu en exécution

Est exécuter seulement lorsque 'state == :ok'

Description

Il assigne 2 variables qui sont:

time --> qui est le temps du système actuellement

dt --> qui est la conversion du temps actuel en tick moins le dernier temps en tick que la partie a été updated.

Par la suite, avec le game qu'il reçois il appel les fonctions privé suivante:

move_bird --> calcul les paramètres de bird (wings, velocity et y).

move_pipes --> calcul la nouvelle position des tuyaux.

check_for_collisions --> Compare la position du personnage(oiseau) et des obstacles(tuyaux, sol et bordure du navigateur web). Puis change le state selon le résultat.

add_pipes --> Ajout des tuyaux a une distance régulière, mais avec une hauteur aléatoire.

Ensuite, il augmente le score de la partie de 0.1 * dt

Finalement, il change le updated de la game pour time

Exemples

iex> update(%{field: value})
{:ok, %Game{}}

iex> update(%{field: bad_value})
{:error, %Ecto.Changeset{}}

Type d'appel

Game.update(game)

Type de retour

retourne game