FlappyBirdWeb.GameLive (FlappyBird v1.0.0)

Module de liaison entre son controller dans flappy_bird et sa vue portant le même nom: game.

Functions

handle_event(binary, arg2, socket)

Attrapeur d'événement

Description

récupère game dans le socket reçus, puis effectue l'action selon le nom de l'événement qui lui est associé.

Nom d'événement disponible

keydown --> selon le state de game

:ok --> appel la fonction Game.flap() et assigne la valeur de retour au game contenue dans le socket

autre --> appel la fonction new_game()

save --> récupère game dans le socket et passe son score convertie en string à la fonction Game.save_score()

Paramètre

binary --> nom de l'événement en string

arg2 --> non utilisé

Socket

mount(params, session, socket)

Constructeur

Description

appel la fonction new_game de FlappyBird.Game avant de faire un rendu de sa vue, en retournant ainsi :ok et game

Paramètre

Params --> non utilisé

Session --> non utilisé

Socket

render(assigns)

Rendu de la page