FlappyBirdWeb.ScoreLive (FlappyBird v1.0.0)

Module de liaison entre son controller dans flappy_bird et sa vue portant le même nom: score.

Functions

mount(params, session, socket)

Constructeur

Description

appel la fonction Scores.subscribe pour être appeller a chaque changement de la liste des scores. Avant de faire un rendu de sa vue, en retournant les scores -> Scores.list_scores()

Paramètre

Params --> non utilisé

Session --> non utilisé

Socket

render(assigns)

Rendu de la page